Module Sha256

SHA256 OCaml binding

type ctx

Context type - opaque.

type buf = (int, Bigarray.int8_unsigned_elt, Bigarray.c_layout) Bigarray.Array1.t

Buffer type.

type t

Digest type - opaque.

val zero : t

The zero digest.

val init : unit -> ctx

Create a new context.

val unsafe_update_substring : ctx -> string -> int -> int -> unit

Sha256.unsafe_update_substring ctx s ofs len updates the context with the substring of s starting at character number ofs and containing len characters. Unsafe: No range checking!

val update_substring : ctx -> string -> int -> int -> unit

Sha256.update_substring ctx s ofs len updates the context with the substring of s starting at character number ofs and containing len characters.

val update_string : ctx -> string -> unit

Sha256.update_string ctx s updates the context with s.

val update_buffer : ctx -> buf -> unit

Sha256.update_buffer ctx a updates the context with a. Runs parallel to other threads if any exist.

val finalize : ctx -> t

Finalize the context and return digest.

val copy : ctx -> ctx

Return a copy of the context.

val string : string -> t

Return the digest of the given string.

val substring : string -> int -> int -> t

Sha256.substring s ofs len returns the digest of the substring of s starting at character number ofs and containing len characters.

val buffer : buf -> t

Return the digest of the given buffer.

val channel : in_channel -> int -> t

If len is nonnegative, Sha256.channel ic len reads len characters from channel ic and returns their digest, or raises End_of_file if end-of-file is reached before len characters are read. If len is negative, Sha256.channel ic len reads all characters from ic until end-of-file is reached and return their digest.

val file : string -> t

Return the digest of the file whose name is given.

val file_fast : string -> t

Return the digest of the file whose name is given using fast C function.

val output : out_channel -> t -> unit

Write a digest on the given output channel.

val input : in_channel -> t

Read a digest from the given input channel.

val to_bin : t -> string

Return a binary representation of the given digest.

val to_hex : t -> string

Return a printable hexadecimal representation of the given digest.

val equal : t -> t -> bool

Returns whether two hashes are equal.

val of_bin : bytes -> t

Sha256.of_bin digest converts the binary representation of a digest to the internal representation of Sha256.t.

val of_hex : string -> t

Sha256.of_hex digest converts the hexadecimal representation of a digest to the internal representation of Sha256.t.